requirements.txt
streamlit==1.31.1
pandas==2.2.2
numpy==1.26.4
plotly==5.18.0
scikit-learn==1.5.1
scikit-surprise==1.1.3